+++ localstack/services/lambda_/invocation/lambda_models.py
@@ -42,31 +42,6 @@ from localstack.utils.strings import long_uid
 
 LOG = logging.getLogger(__name__)
 
-# To add support for a new runtime, just add it here with the accompanying image postfix and to api_utils.RUNTIMES
-IMAGE_MAPPING = {
-    "python3.7": "python:3.7",
-    "python3.8": "python:3.8",
-    "python3.9": "python:3.9",
-    "python3.10": "python:3.10",
-    "python3.11": "python:3.11",
-    "nodejs12.x": "nodejs:12",
-    "nodejs14.x": "nodejs:14",
-    "nodejs16.x": "nodejs:16",
-    "nodejs18.x": "nodejs:18",
-    "ruby2.7": "ruby:2.7",
-    "ruby3.2": "ruby:3.2",
-    "java8": "java:8",
-    "java8.al2": "java:8.al2",
-    "java11": "java:11",
-    "java17": "java:17",
-    "dotnetcore3.1": "dotnet:core3.1",
-    "dotnet6": "dotnet:6",
-    "go1.x": "go:1",
-    "provided": "provided:alami",
-    "provided.al2": "provided:al2",
-}
-SNAP_START_SUPPORTED_RUNTIMES = [Runtime.java11, Runtime.java17]
-
 
 # TODO: maybe we should make this more "transient" by always initializing to Pending and *not* persisting it?
 @dataclasses.dataclass(frozen=True)
